requirements.txt
numpy 
opencv-python
requests
paho-mqtt